Move new method back to camera tools
